requirements.txt
streamlit
pandas
numpy
openai
langchain
beautifulsoup4
scrapy
selenium
plotly
langgraph
tiktoken
python-dotenv
